move shared components to components dir
