+++ vsc-extension/src-views/components/OutputPanel.tsx
