+++ vsc-extension/src-views/input-panel/App.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { RecoilRoot } from "recoil";
-import { ExtensionStateProvider } from "../shared/ExtensionState";
-import { InputPanel } from "../shared/InputPanel";
-import { RecoilSyncWebview } from "../shared/RecoilSyncWebview";
+import { ExtensionStateProvider } from "../components/ExtensionState";
+import { InputPanel } from "../components/InputPanel";
+import { RecoilSyncWebview } from "../components/RecoilSyncWebview";
 
 const App = () => {
   return (
